Adjusted SmartNodeSelector
- Implemented SmartNodeSelector in TypeScript
- Added dummy JSX component as a wrapper around SmartNodeSelector
- Changed file structure
