+++ src/lib/components/SmartNodeSelector/components/Suggestions.css
@@ -29,7 +29,7 @@
     cursor: pointer;
     box-sizing: border-box;
     height: 34px;
-    line-height: 18px;
+    line-height: 1.5;
     white-space: nowrap;
     overflow: hidden;
     text-overflow: ellipsis;
